+++ src/Contracts/Changelog.md
@@ -8,6 +8,7 @@
 - Dropped PHP 8.0 & PHP 8.1 Support
 - Dropped Laravel 9 Support
 - Added the `Stockable` interface
+- Added the `Merchant` interface
 
 ## 3.x Series